app/requirements.txt
numpy==2.3.3
pickle-mixin==1.0.2
streamlit==1.50.0
streamlit-option-menu==0.4.0
scikit-learn==1.6.1
google-generativeai==0.7.1